chore: moved loyalty files
