+++ src/components/features/loyalty/components/tier-info.tsx
