+++ src/components/features/loyalty/components/tier-ruler.tsx
