+++ src/pages/loyalty/[id].tsx
@@ -1,8 +1,8 @@
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next';
 
+import { LoyaltyProgram } from '@/components/features/loyalty/loyalty';
 import { HeadContainer } from '@/components/molecules/head-container';
 import { DashboardTemplate } from '@/components/templates/dashboard';
-import { LoyaltyProgram } from '@/components/templates/loyalty-program/LoyaltyProgram';
 import { query } from '@/constants/queries';
 import { useAuth } from '@/providers/auth';
 import { gqlAnonMethods } from '@/services/hasura/api';
