+++ src/pages/loyalty/credential/[id].tsx
@@ -1,8 +1,8 @@
 import { useRouter } from 'next/router';
 
+import { LoyaltyProgramCredential } from '@/components/features/loyalty/loyalty-credential';
 import { HeadContainer } from '@/components/molecules/head-container';
 import { DashboardTemplate } from '@/components/templates/dashboard';
-import { LoyaltyProgramCredential } from '@/components/templates/loyalty-program/LoyaltyProgramCredential';
 import { query } from '@/constants/queries';
 import { ROUTES } from '@/constants/routes';
 import { useAuth } from '@/providers/auth';
